Support cygwin on Windows
When running `check-ecs-exec.sh` in cygwin on Windows, the REGION
variable contains a trailing carriage return character. The script
terminates with the error message "' doesn't match a supported format."
Removing the trailing carriage return character character allows the
rest of the script to complete normally.

The displayed AWS CLI Version also has a trailing carriage return
character, but this is only a cosmetic issue and doesn't affect the
checks performed by the rest of the script.